Add files via upload
Project 16
(M) is a Support Vector Regression-type approach of your choice (in particular, with one or more kernels of your choice) for univariate regression.
(A1) is an algorithm of the class of active-set methods, cf. also here, applied to either the primal or the dual formulation of the SVR.
(A2) is a general-purpose solver applied to an appropriate formulation of the problem.
No off-the-shelf solver is allowed, save of course for (A2).